Updating doc

+++ docs/mkrst.py
@@ -6,13 +6,16 @@ website of the flopy project.
 
 import os
 
+#print current working directory
+print os.getcwd()
+
 #look through the following subdirectories, and grab all of the
 #modules that should be added to the sphinx documentation.
-flopypth = '../trunk/flopy/'
-pthlist = ['modflow', 'mt3dms', 'seawat', 'utils']
+flopypth = os.path.join('..', '..', '..', 'flopy3.git', 'flopy')
+pthlist = ['modflow', 'mt3d', 'seawat', 'utils']
 namelist = []
 for pth in pthlist:
-    dirpth = flopypth + pth
+    dirpth = os.path.join(flopypth, pth)
     filelist = os.listdir(dirpth)
     for filename in filelist:
         if '.pyc' in filename:
@@ -33,17 +36,18 @@ for name in namelist:
     fnamelist.write('   ' + name + '\n')
     prefix = name.strip().split('.')[2]
     fname = prefix + '.rst'
-    f = open(fname, 'w')
-    s = name + ' Module'
-    f.write(s + '\n')
-    s = '========================'
-    f.write(s + '\n\n')
-    s = '.. automodule:: ' + name
-    f.write(s + '\n')
-    s = '   :members:'
-    f.write(s + '\n')
-    s = '   :inherited-members:'
-    f.write(s + '\n')
-    f.close()
+    if not os.path.exists(fname):
+        f = open(fname, 'w')
+        s = name + ' Module'
+        f.write(s + '\n')
+        s = len(s) * '='
+        f.write(s + '\n\n')
+        s = '.. automodule:: ' + name
+        f.write(s + '\n')
+        s = '   :members:'
+        f.write(s + '\n')
+        s = '   :inherited-members:'
+        f.write(s + '\n')
+        f.close()
     
 fnamelist.close()
